api-gateway/requirements.txt
asgiref==3.7.2
boto3==1.34.54
botocore==1.34.54
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.5
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-gis==1.0
djangorestframework-simplejwt==4.6.0
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
packaging==23.2
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
slack_sdk==3.27.1
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.0.7
